requirements.txt
# Pathway RAG Framework with document parsing support
pathway[xpack-llm-docs]>=0.15.0

# Google Gemini API
google-generativeai>=0.8.0

# LiteLLM for unified LLM interface
litellm>=1.55.0

# Environment variables
python-dotenv>=1.0.0

# Pydantic for configuration
pydantic>=2.0.0

# Unstructured parser dependencies
python-magic>=0.4.27
unstructured[pdf]>=0.10.0

# PDF processing
pdf2image>=1.16.0
pypdf>=3.0.0
pytesseract>=0.3.10
pdfminer.six>=20221105
opencv-python-headless>=4.8.0
